+++ lib/commands/ssl_.ml
@@ -17,6 +17,8 @@ open Lwt
 open Server_config
 open X509.Encoding.Pem
 
+exception PasswordRequired
+
 type keys = Nocrypto.Rsa.pub * Nocrypto.Rsa.priv
 
 let get_pub_key priv =
@@ -30,13 +32,31 @@ let create_cert config =
 let create_user_cert ~cert ~priv_key =
   X509_lwt.private_of_pems ~cert ~priv_key
 
-let get_user_keys ~user config =
+let key_encrypted key =
+  let prefix = "-----BEGIN RSA PRIVATE KEY-----" in
+  (String.length key > String.length prefix &&
+  String.sub key 0 (String.length prefix) = prefix) = false
+
+let get_user_keys ~user ?pswd config =
   let path name =
     let p = Filename.concat config.user_cert_path name in
     Regex.replace ~regx:"%user%" ~tmpl:user p 
   in
-  create_user_cert ~cert:(path config.pem_name) 
-    ~priv_key:(path config.key_name) >>= fun (_,priv) ->
+  Lwt_io.with_file ~mode:Lwt_io.input (path config.key_name) (fun r ->
+    Lwt_io.read r
+  ) >>= fun key ->
+  let key =
+  if key_encrypted key then (
+    match pswd with
+    | Some pswd ->
+      let pswd = Imap_crypto.get_hash_raw (user ^ "\000" ^ pswd) in
+      (Imap_crypto.aes_decrypt_pswd ~pswd key)
+    | None -> raise PasswordRequired
+  ) else ( (* should raise if auth_required and the key is not encrypted? TBD *)
+    key 
+  )
+  in
+  let priv = (PK.of_pem_cstruct1 (Cstruct.of_string key)) in
   return (get_pub_key priv,priv)
 
 let get_system_keys config =